Make SSID display properly
`lv_label_set_text_fmt` needed a `c_str()` instead of a `String`.

Also, I changed `KnownWifi_t.ssid` and `.password` to `.known_ssid` and `.known_password`, to avoid confusion with the class members `ssid` and `password`.

And I changed the class members `ssid` and `password` to `ssid_` and `password_`.