Add watch name to home screen
Just to see how it would look. It might make the screen look too cluttered - but if you have two watches, you want a way to tell them apart. Jan, I did this in a separate commit so it will be easy to back out if you don't want it on the home screen.