DRAFT: Double-tap to switch between LIGHT and DARK themes
It works, but it happens on ALL screens. Any screen with a button (even the "Back" button in the upper left corner) will not look right after a theme change if you do the double-tap on that screen.

We need to either make the double-tap work ONLY on the home screen, or we need a way to update all the buttons and icons on every screen if you double-tap on it to change the theme.