Update graphic elements on double-tap from any screen
This makes the double-tap (which changes from LIGHT to DARK or vice versa) work properly on any screen.

There are also some general code clean-up changes, and a few bug fixes.

Still need to make the double-tap theme change be saved to disk. Now, that works only if you happen to do the double-tap on the Display Settings screen.