First attempt at displaying msgs one at a time
It works for the first message, but crashes when you tap "OK" on the first message to clear it from the screen. It fails in Gui::msg_box_cb() when it tries to call gui->set_display_next_pending_message(true);.
If I comment out that line, then it fails on gui->display_next_message();
I copied Gui::lv_mainbar_callback exactly (I think), but it still crashes.