Implement wifi retry
I couldn't come up with a clever algorithm to get the intervals I wanted, so I implemented it as an array. But it actually works really well, and we *could* add a Settings page to set all of the intervals, if we wanted to. But we can change it from an array to an algorithm if we come up with one.

Still need to eliminate some of the battery-draining impact of all the messages (the vibrating and the display of messages). Will work on that next.